+++ src/features/projects/data/index.ts
@@ -1,7 +1,5 @@
 export { default as GitHubProjectDataSource } from "./GitHubProjectDataSource"
 export * from "./GitHubProjectDataSource"
 export { default as useProjects } from "./useProjects"
-export type { default as IGitHubLoginDataSource } from "./IGitHubLoginDataSource"
 export { default as GitHubLoginDataSource } from "./GitHubLoginDataSource"
 export { default as GitHubRepositoryDataSource } from "./GitHubRepositoryDataSource"
-export * from "./IGitHubGraphQLClient"
