+++ src/features/projects/domain/index.ts
@@ -1,7 +1,10 @@
 export { default as CachingProjectDataSource } from "./CachingProjectDataSource"
 export { default as getSelection } from "./getSelection"
+export type { default as IGitHubLoginDataSource } from "./IGitHubLoginDataSource"
 export type { default as IGitHubRepositoryDataSource } from "./IGitHubRepositoryDataSource"
 export * from "./IGitHubRepositoryDataSource"
+export type { default as IGitHubGraphQLClient } from "./IGitHubGraphQLClient"
+export * from "./IGitHubGraphQLClient"
 export type { default as IProjectConfig } from "./IProjectConfig"
 export * from "./IProjectConfig"
 export type { default as IProjectDataSource } from "./IProjectDataSource"
